requirements.txt
bytewax==0.19
haystack-ai
python-dotenv